Changed Comments
Added to description so it was clear to user which list contained which sublist so that user didn't download the same list twice. Also added new whitelist containing common whitelisted items. Also took out descriptions of list that were not used i.e. b_moab and b_adzhosts.

Also reordered list so that master list was followed by the sub list that are included in the master. Also reordered list so that the ones most likely to give user problems are first and the ones user is most likely to keep are listed last. Also changed default to include steven black by default.